+++ moves.c
@@ -6,7 +6,7 @@
 /*   By: dgutak <dgutak@student.42vienna.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/19 21:20:12 by vfrants           #+#    #+#             */
-/*   Updated: 2023/11/26 15:23:59 by dgutak           ###   ########.fr       */
+/*   Updated: 2023/11/26 18:48:33 by dgutak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	go_forward(t_data *data)
 {
-	const float	delta_x = (data->dir.x - data->player.x) / DIR_L;
-	const float	delta_y = (data->dir.y - data->player.y) / DIR_L;
+	const double	delta_x = (data->dir.x - data->player.x) / DIR_L;
+	const double	delta_y = (data->dir.y - data->player.y) / DIR_L;
 
 	if (data->input.map[(int)data->player.y / IMAGE_SIZE][(int)(data->player.x
 		+ SPEED * delta_x) / IMAGE_SIZE] == '0')
@@ -37,8 +37,8 @@ void	go_forward(t_data *data)
 
 void	go_backward(t_data *data)
 {
-	const float	delta_x = -(data->dir.x - data->player.x) / DIR_L;
-	const float	delta_y = -(data->dir.y - data->player.y) / DIR_L;
+	const double	delta_x = -(data->dir.x - data->player.x) / DIR_L;
+	const double	delta_y = -(data->dir.y - data->player.y) / DIR_L;
 
 	if (data->input.map[(int)data->player.y / IMAGE_SIZE][(int)(data->player.x
 		+ SPEED * delta_x) / IMAGE_SIZE] == '0')
@@ -60,8 +60,8 @@ void	go_backward(t_data *data)
 
 void	go_left(t_data *data)
 {
-	const float	delta_x = (data->dir.y - data->player.y) / DIR_L;
-	const float	delta_y = -(data->dir.x - data->player.x) / DIR_L;
+	const double	delta_x = (data->dir.y - data->player.y) / DIR_L;
+	const double	delta_y = -(data->dir.x - data->player.x) / DIR_L;
 
 	if (data->input.map[(int)data->player.y / IMAGE_SIZE][(int)(data->player.x
 		+ SPEED * delta_x) / IMAGE_SIZE] == '0')
@@ -79,12 +79,13 @@ void	go_left(t_data *data)
 		data->plane.y += SPEED * delta_y;
 		data->plane2.y += SPEED * delta_y;
 	}
+
 }
 
 void	go_right(t_data *data)
 {
-	const float	delta_x = -(data->dir.y - data->player.y) / DIR_L;
-	const float	delta_y = (data->dir.x - data->player.x) / DIR_L;
+	const double	delta_x = -(data->dir.y - data->player.y) / DIR_L;
+	const double	delta_y = (data->dir.x - data->player.x) / DIR_L;
 
 	if (data->input.map[(int)data->player.y / IMAGE_SIZE][(int)(data->player.x
 		+ SPEED * delta_x) / IMAGE_SIZE] == '0')
@@ -102,4 +103,5 @@ void	go_right(t_data *data)
 		data->plane.y += SPEED * delta_y;
 		data->plane2.y += SPEED * delta_y;
 	}
+
 }
